requirements.txt
fastapi
numpy
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
scikit-learn==1.7.2
uvicorn==0.41.0
azure-eventhub
